requirements.txt
google-adk
streamlit
google-generativeai
openai
pandas 
python-dotenv 
matplotlib 
seaborn 
sqlparse
pandasql
openpyxl
sqlalchemy
psycopg2-binary
supabase
fastapi 
uvicorn
plotly
reportlab
markdown
xlsxwriter